Use OpenAPI Generator not Swagger Codegen
This commit replaces [Swagger Codegen][1] with [OpenAPI Generator][2].

Both are code generators for OpenAPI specs. [OpenAPI Generator is a fork
of Swagger CodeGen][3].

Looking at [Open Hub][4] it is clear that [OpenAPI Generator has a more
active community][5] [than Swagger Codegen][6]. And OpenAPI Generator is
a completely open source community.  So OpenAPI Generator it is.

[1]: https://swagger.io/tools/swagger-codegen/
[2]: https://openapi-generator.tech/
[3]: https://openapi-generator.tech/docs/fork-qna
[4]: https://www.openhub.net
[5]: https://www.openhub.net/p/openapi-generator
[6]: https://www.openhub.net/p/swagger-codegen